+++ mycodo/config.py
@@ -14,7 +14,7 @@ sys.path.append(os.path.abspath(os.path.dirname(__file__)))
 from config_translations import TRANSLATIONS
 
 MYCODO_VERSION = '8.8.8'
-ALEMBIC_VERSION = '313a6fb99082'
+ALEMBIC_VERSION = '706a32b64ba2'
 
 #  FORCE_UPGRADE_MASTER
 #  Set True to enable upgrading to the master branch of the Mycodo repository.